#21: Refactor: Push getConfigMap to client mock util test
